requirements.txt
plyer>=2.1.0
pystray>=0.19.5
Pillow>=10.0.0
